feat(support for multiple months): support for multiple months
This commit changes the the calendar return type to 3 dimensional array of dates (Date[][][]). The
first dimention represents the month, the second the week and the thrid the day. This allows you to
creat calendars with multiple months.

BREAKING CHANGE: The return type of calendar has changed from Date[][] to Date[][][]